feat(ui): overhaul main screen UI & add tracker setting
This commit introduces a significant overhaul of the main interface, transitioning to a more structured layout with a functional activity chart and improved tracker lifecycle management.

Key changes include:

    Migrated primary data flow to use getAllVisibleTrackers to support a new soft-delete (hidden) status for trackers.

    Refactored MainActivity layout to utilize movableContentOf, optimizing transitions between portrait (Column) and landscape (Row) orientations.

    Implemented a LazyColumn for tracker lists with custom content padding and a centered add button.

    Replaced basic ID types with UUID strings for Tracker and TimeEvent entities to improve data integrity.

    Added TrackerColor domain model and integrated it into the ActivityChart for color-coded visualization of time segments.

    Introduced a TrackerSettingDialog providing options to edit labels, change tracker colors, or soft-delete trackers.

    Refactored MathUtils to use Double for range mapping, increasing precision for UI coordinate calculations.

    Renamed and cleaned up several UI components (e.g., ActivityCart to ActivityChart, TrackerRunToggleIcon to TrackerRunStatusIcon) for better clarity.

Composed-by: Gemini-3-Flash

+++ app/src/main/kotlin/io/github/aaalest/lanstopwatch/core/utils/MathUtils.kt
@@ -4,12 +4,12 @@ package io.github.aaalest.lanstopwatch.core.utils
 /**
  * Maps a value from one range to another.
  */
-fun Long.mapRange(
-    inMin: Long,
-    inMax: Long,
-    outMin: Long,
-    outMax: Long
-): Long {
+fun Double.mapRange(
+    inMin: Double,
+    inMax: Double,
+    outMin: Double,
+    outMax: Double
+): Double {
     if (inMax == inMin) return outMin
     val rangeSize = (inMax - inMin)
     val valueOffset = (this - inMin)
